+++ packages/vm/src/calls.rs
@@ -1,12 +1,14 @@
 use serde::de::DeserializeOwned;
 use wasmer::Value;
 
-use cosmwasm_std::{ContractResult, CustomMsg, Env, MessageInfo, QueryResponse, Reply, Response};
+use cosmwasm_std::{
+    ContractResult, CustomMsg, Env, IbcBasicResponse, IbcDestinationCallbackMsg,
+    IbcSourceCallbackMsg, MessageInfo, QueryResponse, Reply, Response,
+};
 #[cfg(feature = "stargate")]
 use cosmwasm_std::{
-    Ibc3ChannelOpenResponse, IbcBasicResponse, IbcChannelCloseMsg, IbcChannelConnectMsg,
-    IbcChannelOpenMsg, IbcPacketAckMsg, IbcPacketReceiveMsg, IbcPacketTimeoutMsg,
-    IbcReceiveResponse,
+    Ibc3ChannelOpenResponse, IbcChannelCloseMsg, IbcChannelConnectMsg, IbcChannelOpenMsg,
+    IbcPacketAckMsg, IbcPacketReceiveMsg, IbcPacketTimeoutMsg, IbcReceiveResponse,
 };
 
 use crate::backend::{BackendApi, Querier, Storage};
@@ -54,6 +56,10 @@ mod read_limits {
     /// Max length (in bytes) of the result data from a ibc_packet_timeout call.
     #[cfg(feature = "stargate")]
     pub const RESULT_IBC_PACKET_TIMEOUT: usize = 64 * MI;
+    /// Max length (in bytes) of the result data from a ibc_source_callback call.
+    pub const RESULT_IBC_SOURCE_CALLBACK: usize = 64 * MI;
+    /// Max length (in bytes) of the result data from a ibc_destination_callback call.
+    pub const RESULT_IBC_DESTINATION_CALLBACK: usize = 64 * MI;
 }
 
 /// The limits for the JSON deserialization.
@@ -93,6 +99,10 @@ mod deserialization_limits {
     /// Max length (in bytes) of the result data from a ibc_packet_timeout call.
     #[cfg(feature = "stargate")]
     pub const RESULT_IBC_PACKET_TIMEOUT: usize = 256 * KI;
+    /// Max length (in bytes) of the result data from a ibc_source_callback call.
+    pub const RESULT_IBC_SOURCE_CALLBACK: usize = 256 * KI;
+    /// Max length (in bytes) of the result data from a ibc_destination_callback call.
+    pub const RESULT_IBC_DESTINATION_CALLBACK: usize = 256 * KI;
 }
 
 pub fn call_instantiate<A, S, Q, U>(
@@ -327,6 +337,45 @@ where
     Ok(result)
 }
 
+pub fn call_ibc_source_callback<A, S, Q, U>(
+    instance: &mut Instance<A, S, Q>,
+    env: &Env,
+    msg: &IbcSourceCallbackMsg,
+) -> VmResult<ContractResult<IbcBasicResponse<U>>>
+where
+    A: BackendApi + 'static,
+    S: Storage + 'static,
+    Q: Querier + 'static,
+    U: DeserializeOwned + CustomMsg,
+{
+    let env = to_vec(env)?;
+    let msg = to_vec(msg)?;
+    let data = call_ibc_source_callback_raw(instance, &env, &msg)?;
+    let result = from_slice(&data, deserialization_limits::RESULT_IBC_SOURCE_CALLBACK)?;
+    Ok(result)
+}
+
+pub fn call_ibc_destination_callback<A, S, Q, U>(
+    instance: &mut Instance<A, S, Q>,
+    env: &Env,
+    msg: &IbcDestinationCallbackMsg,
+) -> VmResult<ContractResult<IbcBasicResponse<U>>>
+where
+    A: BackendApi + 'static,
+    S: Storage + 'static,
+    Q: Querier + 'static,
+    U: DeserializeOwned + CustomMsg,
+{
+    let env = to_vec(env)?;
+    let msg = to_vec(msg)?;
+    let data = call_ibc_destination_callback_raw(instance, &env, &msg)?;
+    let result = from_slice(
+        &data,
+        deserialization_limits::RESULT_IBC_DESTINATION_CALLBACK,
+    )?;
+    Ok(result)
+}
+
 /// Calls Wasm export "instantiate" and returns raw data from the contract.
 /// The result is length limited to prevent abuse but otherwise unchecked.
 pub fn call_instantiate_raw<A, S, Q>(
@@ -557,6 +606,44 @@ where
         "ibc_packet_timeout",
         &[env, msg],
         read_limits::RESULT_IBC_PACKET_TIMEOUT,
+    )
+}
+
+pub fn call_ibc_source_callback_raw<A, S, Q>(
+    instance: &mut Instance<A, S, Q>,
+    env: &[u8],
+    msg: &[u8],
+) -> VmResult<Vec<u8>>
+where
+    A: BackendApi + 'static,
+    S: Storage + 'static,
+    Q: Querier + 'static,
+{
+    instance.set_storage_readonly(false);
+    call_raw(
+        instance,
+        "ibc_source_callback",
+        &[env, msg],
+        read_limits::RESULT_IBC_SOURCE_CALLBACK,
+    )
+}
+
+pub fn call_ibc_destination_callback_raw<A, S, Q>(
+    instance: &mut Instance<A, S, Q>,
+    env: &[u8],
+    msg: &[u8],
+) -> VmResult<Vec<u8>>
+where
+    A: BackendApi + 'static,
+    S: Storage + 'static,
+    Q: Querier + 'static,
+{
+    instance.set_storage_readonly(false);
+    call_raw(
+        instance,
+        "ibc_destination_callback",
+        &[env, msg],
+        read_limits::RESULT_IBC_DESTINATION_CALLBACK,
     )
 }
 
@@ -866,10 +953,13 @@ mod tests {
             mock_ibc_packet_ack, mock_ibc_packet_recv, mock_wasmd_attr,
         };
         use cosmwasm_std::{
-            Event, IbcAcknowledgement, IbcOrder, ReplyOn, SubMsgResponse, SubMsgResult,
+            Event, IbcAckCallbackMsg, IbcAcknowledgement, IbcOrder, IbcTimeoutCallbackMsg, ReplyOn,
+            SubMsgResponse, SubMsgResult,
         };
-        static CONTRACT: &[u8] = include_bytes!("../testdata/ibc_reflect.wasm");
+        const CONTRACT: &[u8] = include_bytes!("../testdata/ibc_reflect.wasm");
+        const IBC_CALLBACKS: &[u8] = include_bytes!("../testdata/ibc_callbacks.wasm");
         const IBC_VERSION: &str = "ibc-reflect-v1";
+
         fn setup(
             instance: &mut Instance<MockApi, MockStorage, MockQuerier>,
             channel_id: &str,
@@ -923,13 +1013,16 @@ mod tests {
             };
             call_reply::<_, _, _, Empty>(instance, &mock_env(), &response).unwrap();
         }
+
         const CHANNEL_ID: &str = "channel-123";
         const ACCOUNT: &str = "account-456";
+
         #[test]
         fn call_ibc_channel_open_and_connect_works() {
             let mut instance = mock_instance(CONTRACT, &[]);
             setup(&mut instance, CHANNEL_ID, ACCOUNT);
         }
+
         #[test]
         fn call_ibc_channel_close_works() {
             let mut instance = mock_instance(CONTRACT, &[]);
@@ -941,6 +1034,7 @@ mod tests {
                 .unwrap()
                 .unwrap();
         }
+
         #[test]
         fn call_ibc_packet_ack_works() {
             let mut instance = mock_instance(CONTRACT, &[]);
@@ -951,6 +1045,7 @@ mod tests {
                 .unwrap()
                 .unwrap();
         }
+
         #[test]
         fn call_ibc_packet_timeout_works() {
             let mut instance = mock_instance(CONTRACT, &[]);
@@ -960,6 +1055,7 @@ mod tests {
                 .unwrap()
                 .unwrap();
         }
+
         #[test]
         fn call_ibc_packet_receive_works() {
             let mut instance = mock_instance(CONTRACT, &[]);
@@ -969,6 +1065,65 @@ mod tests {
             call_ibc_packet_receive::<_, _, _, Empty>(&mut instance, &mock_env(), &msg)
                 .unwrap()
                 .unwrap();
+        }
+
+        #[test]
+        fn call_ibc_source_callback_works() {
+            let mut instance = mock_instance(IBC_CALLBACKS, &[]);
+
+            // init
+            let creator = instance.api().addr_make("creator");
+            let info = mock_info(&creator, &[]);
+            call_instantiate::<_, _, _, Empty>(&mut instance, &mock_env(), &info, br#"{}"#)
+                .unwrap()
+                .unwrap();
+
+            /// Response type for the `callback_stats` query
+            #[derive(serde::Serialize, serde::Deserialize)]
+            struct CallbackStats {
+                pub ibc_ack_callbacks: Vec<IbcPacketAckMsg>,
+                pub ibc_timeout_callbacks: Vec<IbcPacketTimeoutMsg>,
+            }
+
+            // send ack callback
+            let ack = mock_ibc_packet_ack(CHANNEL_ID, br#"{}"#, IbcAcknowledgement::new(br#"{}"#))
+                .unwrap();
+            let msg = IbcSourceCallbackMsg::Acknowledgement(IbcAckCallbackMsg::new(
+                ack.acknowledgement,
+                ack.original_packet,
+                ack.relayer,
+            ));
+            call_ibc_source_callback::<_, _, _, Empty>(&mut instance, &mock_env(), &msg)
+                .unwrap()
+                .unwrap();
+            // query the CallbackStats
+            let stats: CallbackStats = serde_json::from_slice(
+                &call_query::<_, _, _>(&mut instance, &mock_env(), br#"{"callback_stats":{}}"#)
+                    .unwrap()
+                    .unwrap(),
+            )
+            .unwrap();
+            assert_eq!(1, stats.ibc_ack_callbacks.len());
+            assert_eq!(0, stats.ibc_timeout_callbacks.len());
+
+            // send timeout callback
+            let timeout = mock_ibc_packet_timeout(CHANNEL_ID, br#"{}"#).unwrap();
+            let msg = IbcSourceCallbackMsg::Timeout(IbcTimeoutCallbackMsg::new(
+                timeout.packet,
+                timeout.relayer,
+            ));
+            call_ibc_source_callback::<_, _, _, Empty>(&mut instance, &mock_env(), &msg)
+                .unwrap()
+                .unwrap();
+            // query the CallbackStats
+            let stats: CallbackStats = serde_json::from_slice(
+                &call_query::<_, _, _>(&mut instance, &mock_env(), br#"{"callback_stats":{}}"#)
+                    .unwrap()
+                    .unwrap(),
+            )
+            .unwrap();
+            assert_eq!(1, stats.ibc_ack_callbacks.len());
+            assert_eq!(1, stats.ibc_timeout_callbacks.len());
         }
     }
 }